ml-api/requirements.txt
Flask==2.1.0
torch==2.2.0
Pillow==11.0.0